convert tables 6-77 and 6-78 (DAC LR volume)
The datasheet gives tables of register values for DAC Left and DAC
Right volume using a bunch of base-2 constants. This code converts
that stuff to a table of decimal values, then tests a dB to int8
conversion function. You can compare the table to the output of the
conversion function to see that they match and that out of range
values get appropriately clipped.