script to approximate Table 6-24
The volume table is a piecewise non-linear function. The bulk of
it can be accurately approximated with the formula

    register_value = round((-1.99 * gain_dB) - 0.2)

to compute register values as a function of gain dB in the range of
-52.7 dB <= gain_dB <= 0 dB.

The table values for gain_dB < -52.7 are not a straight line, so
the code just uses a lookup table.